requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-markdown-view==0.0.5
django-nested-admin==4.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.7
inflection==0.5.1
Markdown==3.6
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
#psycopg2==2.9.9 #instalado direto no SO
#psycopg2-binary==2.9.9
PyJWT==2.8.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
faker